requirements.txt
flask
flask-cors
tensorflow>=2.10
tensorflow-hub
pillow
numpy
gunicorn